+++ packages/util/src/lib/functions/specific-date.function.ts
@@ -14,6 +14,43 @@ export function specificDateToNumber(date: SpecificDate): number {
     return date.day + (((seasonMap.get(date.season)) ?? 0) - 1) * 28 + (date.year - 1) * 112
 }
 
+export function dateInRanges(date: SpecificDate, range: { from: SpecificDate, to: SpecificDate } | {
+    from: SpecificDate,
+    to: SpecificDate
+}[], ignoreYear = false): boolean {
+
+    const rangeAsList = [range].flat();
+
+    if (!rangeAsList.length) return false;
+
+    let dateValue = specificDateToNumber(date);
+
+    if (ignoreYear) {
+        dateValue = Math.floor(dateValue - 1) % 112;
+    }
+
+    return rangeAsList.some(r => {
+        let start = specificDateToNumber(r.from);
+
+        if (ignoreYear) {
+            start = Math.floor(start - 1) % 112;
+        }
+
+
+        let end = specificDateToNumber(r.to);
+
+
+        if (ignoreYear) {
+            end = Math.floor(end - 1) % 112;
+        }
+
+        return start <= dateValue && end >= dateValue;
+
+    })
+
+
+}
+
 export function numberToSpecificDate(days: number): SpecificDate {
 
     const year = Math.floor((days - 1) / 112) + 1;